End5SequenceData/End3SequenceData/EndSequenceData

The End5SequenceData/End3SequenceData/EndSequenceData classes aggregate the counts of read ends at each position along a transcript. End5SequenceData/End3SequenceData classes aggregate either the 5'-end or 3'-end, the EndSequenceData aggregates both.

All three classes contain one column per data file named using the following naming convention (end5/end3/end).condition.replicate.

aggregate calculates the mean and sd for samples in the control and treated condition separatly.

End5SequenceDataFrame(
  df,
  ranges,
  sequence,
  replicate,
  condition,
  bamfiles,
  seqinfo
)

End3SequenceDataFrame(
  df,
  ranges,
  sequence,
  replicate,
  condition,
  bamfiles,
  seqinfo
)

EndSequenceDataFrame(
  df,
  ranges,
  sequence,
  replicate,
  condition,
  bamfiles,
  seqinfo
)

End5SequenceData(bamfiles, annotation, sequences, seqinfo, ...)

End3SequenceData(bamfiles, annotation, sequences, seqinfo, ...)

EndSequenceData(bamfiles, annotation, sequences, seqinfo, ...)

# S4 method for End5SequenceData,BamFileList,GRangesList,XStringSet,ScanBamParam
getData(x, bamfiles, grl, sequences, param, args)

# S4 method for End3SequenceData,BamFileList,GRangesList,XStringSet,ScanBamParam
getData(x, bamfiles, grl, sequences, param, args)

# S4 method for EndSequenceData,BamFileList,GRangesList,XStringSet,ScanBamParam
getData(x, bamfiles, grl, sequences, param, args)

# S4 method for End5SequenceData
aggregateData(x, condition = c("Both", "Treated", "Control"))

# S4 method for End3SequenceData
aggregateData(x, condition = c("Both", "Treated", "Control"))

# S4 method for EndSequenceData
aggregateData(x, condition = c("Both", "Treated", "Control"))

# S4 method for EndSequenceData
getDataTrack(x, name, ...)

# S4 method for End5SequenceData
getDataTrack(x, name, ...)

# S4 method for End3SequenceData
getDataTrack(x, name, ...)

Arguments

df, ranges, sequence, replicate

inputs for creating a SequenceDataFrame. See SequenceDataFrame.

condition

For aggregate: condition for which the data should be aggregated.

bamfiles, annotation, seqinfo, grl, sequences, param, args, ...

See SequenceData and SequenceData-functions

x

a End5SequenceData, End3SequenceData or EndSequenceData object

name

For getDataTrack: a valid transcript name. Must be a name of ranges(x).

Value

a End5SequenceData, a End3SequenceData or a EndSequenceData object

Examples

# Construction of a End5SequenceData object
library(RNAmodR.Data)
library(rtracklayer)
annotation <- GFF3File(RNAmodR.Data.example.man.gff3())
#> see ?RNAmodR.Data and browseVignettes('RNAmodR.Data') for documentation
#> loading from cache
sequences <- RNAmodR.Data.example.man.fasta()
#> see ?RNAmodR.Data and browseVignettes('RNAmodR.Data') for documentation
#> loading from cache
files <- c(treated = RNAmodR.Data.example.wt.1())
#> see ?RNAmodR.Data and browseVignettes('RNAmodR.Data') for documentation
#> loading from cache
e5sd <- End5SequenceData(files, annotation = annotation,
                        sequences = sequences)
#> Warning: makeTxDbFromGFF() has moved to the txdbmaker package. Please call
#>   txdbmaker::makeTxDbFromGFF() to get rid of this warning.
#> Import genomic features from the file as a GRanges object ... 
#> OK
#> Prepare the 'metadata' data frame ... 
#> OK
#> Make the TxDb object ... 
#> Warning: Direct call of 'as.data.frame.factor()' is deprecated.  Use 'as.data.frame.vector()' or 'as.data.frame()' instead
#> Warning: Direct call of 'as.data.frame.factor()' is deprecated.  Use 'as.data.frame.vector()' or 'as.data.frame()' instead
#> Warning: Direct call of 'as.data.frame.factor()' is deprecated.  Use 'as.data.frame.vector()' or 'as.data.frame()' instead
#> Warning: Direct call of 'as.data.frame.factor()' is deprecated.  Use 'as.data.frame.vector()' or 'as.data.frame()' instead
#> Warning: Direct call of 'as.data.frame.factor()' is deprecated.  Use 'as.data.frame.vector()' or 'as.data.frame()' instead
#> Warning: Direct call of 'as.data.frame.factor()' is deprecated.  Use 'as.data.frame.vector()' or 'as.data.frame()' instead
#> Warning: Direct call of 'as.data.frame.factor()' is deprecated.  Use 'as.data.frame.vector()' or 'as.data.frame()' instead
#> Warning: Direct call of 'as.data.frame.factor()' is deprecated.  Use 'as.data.frame.vector()' or 'as.data.frame()' instead
#> OK
#> Loading 5'-end position data from BAM files ... 
#> OK