ModInosine

Inosine can be detected in RNA-Seq data by the conversion of A positions to G. This conversion is detected by ModInosine and used to search for Inosine positions. dataType is "PileupSequenceData".

Only samples labeled with the condition treated are used for this analysis, since the A to G conversion is common feature among the reverse transcriptases usually emploied. Let us know, if that is not the case, and the class needs to be modified.

Further information on Functions of ModInosine.

ModInosine(x, annotation, sequences, seqinfo, ...)

ModSetInosine(x, annotation = NA, sequences = NA, seqinfo = NA, ...)

Arguments

x
the input which can be of the different types depending on whether a ModRiboMethSeq or a ModSetRiboMethSeq object is to be constructed. For more information have a look at the documentation of the Modifier and ModifierSet classes.

annotation
annotation data, which must match the information contained in the BAM files. This is parameter is only required, if x is not a Modifier object.

sequences
sequences matching the target sequences the reads were mapped onto. This must match the information contained in the BAM files. This is parameter is only required, if x is not a Modifier object.

seqinfo
An optional Seqinfo argument or character vector, which can be coerced to one, to subset the sequences to be analyzed on a per chromosome basis.

...
Optional arguments overwriting default values, which are

minCoverage: The minimal coverage at the position as integer value (default: minCoverage = 10L).

minReplicate: minimum number of replicates needed for the analysis (default: minReplicate = 1L).

minScore: minimum score to identify Inosine positions de novo (default: minScore = 0.4).

Value

a ModInosine or ModSetInosine object

Details

ModInosine score: the scores for reported Inosine positions are between 0 and 1. They are calculated as the relative amount of called G bases ((G / N)) and only saved for genomic A positions.

Author

Felix G.M. Ernst [aut]

Examples

# construction of ModInosine object
library(RNAmodR.Data)
library(rtracklayer)
annotation <- GFF3File(RNAmodR.Data.example.man.gff3())

#> snapshotDate(): 2021-01-12

#> see ?RNAmodR.Data and browseVignettes('RNAmodR.Data') for documentation

#> loading from cache

sequences <- RNAmodR.Data.example.man.fasta()

#> snapshotDate(): 2021-01-12

#> see ?RNAmodR.Data and browseVignettes('RNAmodR.Data') for documentation

#> loading from cache

files <- c(treated = RNAmodR.Data.example.wt.1())

#> snapshotDate(): 2021-01-12

#> see ?RNAmodR.Data and browseVignettes('RNAmodR.Data') for documentation

#> loading from cache

mi <- ModInosine(files,annotation = annotation ,sequences = sequences)

#> Import genomic features from the file as a GRanges object ... 

#> OK

#> Prepare the 'metadata' data frame ... 

#> OK

#> Make the TxDb object ... 

#> OK

#> Loading Pileup data from BAM files ... 

#> OK

#> Aggregating data and calculating scores ... 

#> Starting to search for 'Inosine' ... 

#> done.

# construction of ModSetInosine object
if (FALSE) {
files <- list("SampleSet1" = c(treated = RNAmodR.Data.example.wt.1(),
                               treated = RNAmodR.Data.example.wt.2(),
                               treated = RNAmodR.Data.example.wt.3()),
              "SampleSet2" = c(treated = RNAmodR.Data.example.bud23.1(),
                               treated = RNAmodR.Data.example.bud23.2()),
              "SampleSet3" = c(treated = RNAmodR.Data.example.trm8.1(),
                               treated = RNAmodR.Data.example.trm8.2()))
msi <- ModSetInosine(files, annotation = annotation, sequences = sequences)
}