Functions for ModInosine

All of the functions of Modifier and the ModifierSet classes are inherited by the ModInosine and ModSetInosine classes.

Check below for the specifically implemented functions.

# S4 method for ModInosine
settings(x) <- value

# S4 method for ModInosine
aggregateData(x)

# S4 method for ModInosine
findMod(x)

# S4 method for ModInosine
getDataTrack(x, name, type, ...)

# S4 method for ModInosine,GRanges
plotDataByCoord(x, coord, type = "score", window.size = 15L, ...)

# S4 method for ModInosine
plotData(x, name, from = 1L, to = 30L, type = "score", ...)

# S4 method for ModSetInosine,GRanges
plotDataByCoord(x, coord, type = "score", window.size = 15L, ...)

# S4 method for ModSetInosine
plotData(x, name, from = 1L, to = 30L, type = "score", ...)

Arguments

x
a Modifier or a ModifierSet object. For more details see also the man pages for the functions mentioned below.

value
See settings

coord, name, from, to, type, window.size, ...
See plotData

Value

settings See settings.

aggregate See aggregate.

modify See modify.

getDataTrack a list of DataTrack objects. See plotDataByCoord.

plotData See plotDataByCoord.

plotDataByCoord See plotDataByCoord.

Details

ModInosine specific arguments for plotData:

colour.bases - a named character vector of length = 4 for the colours of the individual bases. The names are expected to be c("G","A","U","C")

Examples

data(msi,package="RNAmodR")
mi <- msi[[1]]
settings(mi)

#> $minCoverage
#> [1] 10
#> 
#> $minReplicate
#> [1] 1
#> 
#> $find.mod
#> [1] TRUE
#> 
#> $minScore
#> [1] 0.4
#> 

if (FALSE) {
aggregate(mi)
modify(mi)
}
getDataTrack(mi, "1", mainScore(mi))

#> $score
#> DataTrack 'Score Inosine'
#> | genome: NA
#> | active chromosome: chr1
#> | positions: 1800
#> | samples:1
#> | strand: + 
#>