SequenceData/SequenceDataSet/SequenceDataList/SequenceDataFrame functions

The SequenceData, SequenceDataSet, SequenceDataList and SequenceDataFrame classes share functionality. Have a look at the elements listed directly below.

replicates(x)

# S4 method for SequenceDataFrame
show(object)

# S4 method for SequenceDataFrame
conditions(object)

# S4 method for SequenceDataFrame
bamfiles(x)

# S4 method for SequenceDataFrame
dataType(x)

# S4 method for SequenceDataFrame
ranges(x)

# S4 method for SequenceDataFrame
replicates(x)

# S4 method for SequenceDataFrame
seqinfo(x)

# S4 method for SequenceDataFrame
seqinfo(x)

# S4 method for SequenceDataFrame
seqtype(x)

# S4 method for SequenceDataFrame
seqtype(x) <- value

# S4 method for SequenceDataFrame
sequences(x)

# S4 method for SequenceData
show(object)

# S4 method for SequenceData,BamFileList,GRangesList,XStringSet,ScanBamParam
getData(x, bamfiles, grl, sequences, param, args)

# S4 method for SequenceData
bamfiles(x)

# S4 method for SequenceData
conditions(object)

# S4 method for SequenceData
ranges(x)

# S4 method for SequenceData
replicates(x)

# S4 method for SequenceData
seqinfo(x)

# S4 method for SequenceData
sequences(x)

# S4 method for SequenceData
seqtype(x)

# S4 method for SequenceData
seqtype(x) <- value

# S4 method for SequenceData
dataType(x)

# S4 method for SequenceDataSet
show(object)

# S4 method for SequenceDataSet
bamfiles(x)

# S4 method for SequenceDataSet
conditions(object)

# S4 method for SequenceDataSet
names(x)

# S4 method for SequenceDataSet
ranges(x)

# S4 method for SequenceDataSet
replicates(x)

# S4 method for SequenceDataSet
seqinfo(x)

# S4 method for SequenceDataSet
seqtype(x)

# S4 method for SequenceDataSet
seqtype(x) <- value

# S4 method for SequenceDataSet
sequences(x)

# S4 method for SequenceDataList
show(object)

# S4 method for SequenceDataList
bamfiles(x)

# S4 method for SequenceDataList
conditions(object)

# S4 method for SequenceDataList
names(x)

# S4 method for SequenceDataList
ranges(x)

# S4 method for SequenceDataList
replicates(x)

# S4 method for SequenceDataList
seqinfo(x)

# S4 method for SequenceDataList
seqtype(x)

# S4 method for SequenceDataList
seqtype(x) <- value

# S4 method for SequenceDataList
sequences(x)

Arguments

x, object

a SequenceData, SequenceDataSet, SequenceDataList or a SequenceDataFrame object.

value

a new seqtype, either "RNA" or "DNA"

bamfiles

a BamFileList.

grl

a GRangesList from exonsBy(..., by = "tx")

sequences

a XStringSet of type RNAStringSet, ModRNAStringSet, DNAStringSet or ModDNAStringSet

param

a ScanBamParam object

args

a list of addition arguments

Value

seqinfo: a Seqinfo object ().

sequences: a RNAStingSet object or a RNAString object for a SequenceDataFrame.

ranges: a GRangesList object with each element per transcript or a GRanges object for a SequenceDataFrame.

bamfiles: a BamFileList object or a SimpleList of BamFileList objects for a SequenceDataList.

Examples

data(e5sd,package="RNAmodR")
# general accessors
seqinfo(e5sd)
#> Seqinfo object with 11 sequences from an unspecified genome; no seqlengths:
#>   seqnames seqlengths isCircular genome
#>   chr1             NA         NA   <NA>
#>   chr2             NA         NA   <NA>
#>   chr3             NA         NA   <NA>
#>   chr4             NA         NA   <NA>
#>   chr5             NA         NA   <NA>
#>   chr6             NA         NA   <NA>
#>   chr7             NA         NA   <NA>
#>   chr8             NA         NA   <NA>
#>   chr9             NA         NA   <NA>
#>   chr10            NA         NA   <NA>
#>   chr11            NA         NA   <NA>
sequences(e5sd)
#> RNAStringSet object of length 2:
#>     width seq                                               names               
#> [1]   100 UAUCUGGUUGAUCCUGCCAGUAG...ACAGUGAAACUGCGAAUGGCUCA 1
#> [2]   101 CCGAGAGGUCUUGGUAAUCUUGU...UUCCUAGUAAGCGCAAGUCAUCA 2
ranges(e5sd)
#> GRangesList object of length 2:
#> $`1`
#> GRanges object with 1 range and 3 metadata columns:
#>       seqnames    ranges strand |   exon_id   exon_name exon_rank
#>          <Rle> <IRanges>  <Rle> | <integer> <character> <integer>
#>   [1]     chr1     1-100      + |         1        rna1         1
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths
#> 
#> $`2`
#> GRanges object with 1 range and 3 metadata columns:
#>       seqnames    ranges strand |   exon_id   exon_name exon_rank
#>          <Rle> <IRanges>  <Rle> | <integer> <character> <integer>
#>   [1]     chr1 1500-1600      + |         2        rna2         1
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths
#> 
bamfiles(e5sd)
#> BamFileList of length 3
#> names(3): treated treated treated