The SequenceDataFrame class

The SequenceDataFrame class is a virtual class and contains data for positions along a single transcript. In addition to being used for returning elements from a SequenceData object, the SequenceDataFrame class is used to store the unlisted data within a SequenceData object. Therefore, a matching SequenceData and SequenceDataFrame class must be implemented.

The SequenceDataFrame class is derived from the DataFrame class. To follow the functionallity in the S4Vectors package, SequenceDataFrame implements the concept, whereas SequenceDFrame is the implementation for in-memory data representation from which some specific *SequenceDataFrame class derive from, e.g. CoverageSequenceData.

Subsetting of a SequenceDataFrame returns a SequenceDataFrame or DataFrame, if it is subset by a column or row, respectively. The drop argument is ignored for column subsetting.

# S4 method for SequenceDataFrame
cbind(..., deparse.level = 1)

# S4 method for SequenceDataFrame,ANY,ANY,ANY
[(x, i, j, ..., drop = TRUE)

Arguments

x, i, j, ..., drop, deparse.level

arguments used for subsetting or base::cbind.

Value

A SequenceDataFrame object or if subset to row a DataFrame

Slots

ranges

a GRanges object each element describing a transcript including its element. The GRanges is constructed from the unlisted results of the exonsBy(x, by="tx") function. If during construction a GRangesList is provided instead of a character value pointing to a gff3 file or a TxDb object, it must have a comparable structure.

sequence

a XString of type sequencesType from the parent SequenceData object.

condition

conditions along the BamFileList: Either control or treated

replicate

replicate number along the BamFileList for each of the condition types.

bamfiles

the input bam files as BamFileList

seqinfo

a Seqinfo describing the avialable/used chromosomes.

See also

Examples

data(e5sd,package="RNAmodR")
# A SequenceDataFrame can is usually constructed by subsetting from 
# a SequenceData object
sdf <- e5sd[[1]]
# Its also used to store the unlisted data in a SequenceData object
sdf <- unlist(e5sd) # should probably only used internally
e5sd <- relist(sdf,e5sd)