Settings for Modifier objects

Depending on data prepation, quality and desired stringency of a modification strategy, settings for cut off parameters or other variables may need to be adjusted. This should be rarely the case, but a function for changing these settings, is implemented as the... settings function.

For changing values the input can be either a list or something coercible to a list. Upon changing a setting, the validity of the value in terms of type(!) and dimensions will be checked.

If settings have been modified after the data was loaded, the data is potentially invalid. To update the data, run the aggregate or the modify function.

settings(x, name = NULL)

settings(x, name) <- value

# S4 method for Modifier
settings(x, name = NULL)

# S4 method for Modifier
settings(x) <- value

# S4 method for ModifierSet
settings(x, name = NULL)

# S4 method for ModifierSet
settings(x) <- value

Arguments

x

a Modifier or ModifierSet class

name

name of the setting to be returned or set

value

value of the setting to be set

Value

If name is omitted, settings returns a list of all settings. If name is set, settings returns a single settings or NULL, if a value for name is not available.

Examples

data(msi,package="RNAmodR")
mi <- msi[[1]]
# returns a list of all settings
settings(mi)
#> $minCoverage
#> [1] 10
#> 
#> $minReplicate
#> [1] 1
#> 
#> $find.mod
#> [1] TRUE
#> 
#> $minScore
#> [1] 0.4
#> 
# accesses a specific setting
settings(mi,"minCoverage")
#> [1] 10
# modification of setting
settings(mi) <- list(minCoverage = 11L)