Retrieving information about used reads in RNAmodR

stats returns information about reads used in the RNAmodR analysis. Three modes are available depending on which type of object is provided. If a SequenceData object is provided, a BamFile or BamFileList must be provided as well. If a Modifier object is used, the bam files returned from the bamfiles function are used. This is also the case, if a ModifierSet object is used.

stats(x, file, ...)

# S4 method for SequenceData,BamFile
stats(x, file, ...)

# S4 method for SequenceData,BamFileList
stats(x, file, ...)

# S4 method for Modifier,missing
stats(x)

# S4 method for ModifierSet,missing
stats(x)

Arguments

x
a SequenceData, Modifier or ModifierSet object

file
a BamFile or BamFileList, if x is a SequenceData object.

...
optional parameters used as stated here (except minQuality), if x is a SequenceData object.

Value

a DataFrame, DataFrameList or SimpleList with the results in aggregated form

Examples

library(RNAmodR.Data)
library(rtracklayer)
sequences <- RNAmodR.Data.example.AAS.fasta()

#> snapshotDate(): 2021-01-11

#> see ?RNAmodR.Data and browseVignettes('RNAmodR.Data') for documentation

#> downloading 2 resources

#> retrieving 2 resources

#> loading from cache

annotation <- GFF3File(RNAmodR.Data.example.AAS.gff3())

#> snapshotDate(): 2021-01-11

#> see ?RNAmodR.Data and browseVignettes('RNAmodR.Data') for documentation

#> downloading 1 resources

#> retrieving 1 resource

#> loading from cache

files <- list("SampleSet1" = c(treated = RNAmodR.Data.example.wt.1(),
                               treated = RNAmodR.Data.example.wt.2(),
                               treated = RNAmodR.Data.example.wt.3()),
              "SampleSet2" = c(treated = RNAmodR.Data.example.bud23.1(),
                               treated = RNAmodR.Data.example.bud23.2()),
              "SampleSet3" = c(treated = RNAmodR.Data.example.trm8.1(),
                               treated = RNAmodR.Data.example.trm8.2()))

#> snapshotDate(): 2021-01-11

#> see ?RNAmodR.Data and browseVignettes('RNAmodR.Data') for documentation

#> loading from cache

#> snapshotDate(): 2021-01-11

#> see ?RNAmodR.Data and browseVignettes('RNAmodR.Data') for documentation

#> downloading 2 resources

#> retrieving 2 resources

#> loading from cache

#> snapshotDate(): 2021-01-11

#> see ?RNAmodR.Data and browseVignettes('RNAmodR.Data') for documentation

#> downloading 2 resources

#> retrieving 2 resources

#> loading from cache

#> snapshotDate(): 2021-01-11

#> see ?RNAmodR.Data and browseVignettes('RNAmodR.Data') for documentation

#> downloading 2 resources

#> retrieving 2 resources

#> loading from cache

#> snapshotDate(): 2021-01-11

#> see ?RNAmodR.Data and browseVignettes('RNAmodR.Data') for documentation

#> downloading 2 resources

#> retrieving 2 resources

#> loading from cache

#> snapshotDate(): 2021-01-11

#> see ?RNAmodR.Data and browseVignettes('RNAmodR.Data') for documentation

#> downloading 2 resources

#> retrieving 2 resources

#> loading from cache

#> snapshotDate(): 2021-01-11

#> see ?RNAmodR.Data and browseVignettes('RNAmodR.Data') for documentation

#> downloading 2 resources

#> retrieving 2 resources

#> loading from cache

msi <- ModSetInosine(files, annotation = annotation, sequences = sequences)

#> Import genomic features from the file as a GRanges object ... 

#> OK

#> Prepare the 'metadata' data frame ... 

#> OK

#> Make the TxDb object ... 

#> OK

#> 1. ModInosine analysis 'SampleSet1':

#> Loading Pileup data from BAM files ... 

#> OK

#> Aggregating data and calculating scores ... 

#> Starting to search for 'Inosine' ... 

#> done.

#> 2. ModInosine analysis 'SampleSet2':

#> Loading Pileup data from BAM files ... 

#> OK

#> Aggregating data and calculating scores ... 

#> Starting to search for 'Inosine' ... 

#> done.

#> 3. ModInosine analysis 'SampleSet3':

#> Loading Pileup data from BAM files ... 

#> OK

#> Aggregating data and calculating scores ... 

#> Starting to search for 'Inosine' ... 

#> done.

# smallest chunk of information
stats(sequenceData(msi[[1L]]),bamfiles(msi[[1L]])[[1L]])

#> DataFrame with 11 rows and 6 columns
#>    seqnames seqlength    mapped  unmapped          used     used_distro
#>    <factor> <integer> <numeric> <numeric> <IntegerList>          <List>
#> 1     chr1       1800    197050         0        159782 83,1252,860,...
#> 2     chr2         85      5863         0          2459     2,16,16,...
#> 3     chr3         76     76905         0         63497 35,478,4106,...
#> 4     chr4         77      8299         0          6554     6,27,36,...
#> 5     chr5         74     11758         0          8818  520,105,93,...
#> 6     chr6         76      3504         0          2427     1,38,45,...
#> 7     chr7         75     14556         0         12243  1532,46,12,...
#> 8     chr8         75    144293         0        143068    14,44,48,...
#> 9     chr9         75     13790         0          9753     1,49,43,...
#> 10    chr10        85     19861         0         17729    35,21,10,...
#> 11    chr11        77     10532         0          9086   53,92,185,...

# partial information
stats(sequenceData(msi[[1L]]),bamfiles(msi[[1L]]))

#> DataFrameList of length 3
#> names(3): treated treated treated

# the whole stats
stats(msi)

#> List of length 3
#> names(3): SampleSet1 SampleSet2 SampleSet3